pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "pyyaml>=6.0.2",
]


[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "E", "W", "C90", "I"]
